Initial Set Up For DeepEvals (#1)
Set up PyTest framework to execute DeepEvals, an evals framework to evaluate the output of an LLM model. The test demonstrate LLM As A Judge where it compares the output of a model (ollama 3.1:8b) and uses gpt-4o-mini as a judge.

A simple html viewer is provided to view the results of the test from the jsonl file generated.